Update synchronization.
This change revert to using a sync mutex for coordinating access to pointers and eventually to the buffers.

This will risk blocking on the both sides, but it is the safer option, as I don't feel confident in the atomic ordering and the timing for this particular case right now.